fix: Stop using shlex. Send cmd as list of params.
We had been generating cmd arguments to qri by passing a command string
to shlex. This is usually fine, but if there's any quirk of shell
escaping related to quotes and/or spaces, the user would need to be
aware of it, which doesn't make sense since it's supposed to be under
the hood. An obvious example is if we're executing an sql query with
quoted spaces. (i.e., ... where wbp.country_name="Cayman Islands").

Instead, we can skip the middle man by just splitting the command into
the list of arguments that we already know is supposed to be passed into
qri's CLI.